CHG: Rename submodule to connectivity
- why didn't we do this from the beginning..?

Changes to be committed:
	modified:   syncopy/__init__.py
	renamed:    syncopy/nwanalysis/AV_compRoutines.py -> syncopy/connectivity/AV_compRoutines.py
	renamed:    syncopy/nwanalysis/ST_compRoutines.py -> syncopy/connectivity/ST_compRoutines.py
	renamed:    syncopy/nwanalysis/__init__.py -> syncopy/connectivity/__init__.py
	renamed:    syncopy/nwanalysis/connectivity_analysis.py -> syncopy/connectivity/connectivity_analysis.py
	renamed:    syncopy/nwanalysis/csd.py -> syncopy/connectivity/csd.py
	renamed:    syncopy/nwanalysis/granger.py -> syncopy/connectivity/granger.py
	renamed:    syncopy/nwanalysis/wilson_sf.py -> syncopy/connectivity/wilson_sf.py
	modified:   syncopy/tests/test_connectivity.py
