+++ syncopy/connectivity/ST_compRoutines.py
