+++ syncopy/connectivity/__init__.py
