+++ syncopy/connectivity/csd.py
