+++ syncopy/connectivity/granger.py
