+++ syncopy/tests/test_connectivity.py
@@ -15,7 +15,7 @@ import dask.distributed as dd
 
 import syncopy as spy
 from syncopy import AnalogData, SpectralData
-from syncopy.nwanalysis.connectivity_analysis import connectivity_outputs
+from syncopy.connectivity.connectivity_analysis import connectivity_outputs
 from syncopy import connectivityanalysis as cafunc
 import syncopy.tests.synth_data as synth_data
 import syncopy.tests.helpers as helpers
@@ -180,10 +180,10 @@ class TestGranger:
 
             # Test that 'metadata_keys' in the Granger ComputationalRoutine is up-to-date. All listed
             #  keys should exist...
-            for k in spy.nwanalysis.AV_compRoutines.GrangerCausality.metadata_keys:
+            for k in spy.connectivity.AV_compRoutines.GrangerCausality.metadata_keys:
                 assert k in Gcaus.info
             # ... and no unmentioned extra keys should be in there.
-            assert len(Gcaus.info) == len(spy.nwanalysis.AV_compRoutines.GrangerCausality.metadata_keys)
+            assert len(Gcaus.info) == len(spy.connectivity.AV_compRoutines.GrangerCausality.metadata_keys)
 
     def test_gr_selections(self):
